% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test9-package.R
\docType{package}
\name{test9-package}
\alias{test9}
\alias{test9-package}
\title{test9: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Ramnath Vaidyanathan \email{ramnath.vaidya@gmail.com}

}
\keyword{internal}
